+++ CHANGES.md
@@ -10,13 +10,21 @@
 - **irmin**:
   - Require OCaml 4.07 (#961, @CraigFe)
   - Add sanity checks when creating `Irmin.Type` records, variants and enums
-    (#956, @liautaud):
+    (#956 and #966, @liautaud):
      - `Irmin.Type.sealr` will now raise `Invalid_argument` if two fields of
        the record have the same name;
      - `Irmin.Type.sealv` will now raise `Invalid_argument` if two cases of
        the variant with the same arity have the same name;
      - `Irmin.Type.enum` will now raise `Invalid_argument` if two cases of
        the enum have the same name.
+     - `Irmin.Type.field` will now raise `Invalid_argument` if the name of
+       the field is not a valid UTF-8 string.
+     - `Irmin.Type.case0` will now raise `Invalid_argument` if the name of
+       the case is not a valid UTF-8 string.
+     - `Irmin.Type.case1` will now raise `Invalid_argument` if the name of
+       the case is not a valid UTF-8 string.
+     - `Irmin.Type.enum` will now raise `Invalid_argument` if the name of
+       any of the tags is not a valid UTF-8 string.
   - Changed the JSON encoding of options and unit to avoid ambiguous cases
     (#967, @liautaud):
     - `()` is now encoded as `{}`;